Rename module zonemaintainer to zonemaintenance.
