+++ src/lib.rs
@@ -171,5 +171,5 @@ pub mod utils;
 pub mod validate;
 pub mod validator;
 pub mod zonefile;
-pub mod zonemaintainer;
+pub mod zonemaintainance;
 pub mod zonetree;
